+++ x/evm/keeper/msg_server.go
@@ -350,6 +350,9 @@ func (server msgServer) Send(goCtx context.Context, msg *types.MsgSend) (*types.
 
 func (server msgServer) RegisterPointer(goCtx context.Context, msg *types.MsgRegisterPointer) (*types.MsgRegisterPointerResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
+	if server.GetParams(ctx).RegisterPointerDisabled {
+		return nil, fmt.Errorf("registering CW->ERC pointers has been disabled")
+	}
 	var existingPointer sdk.AccAddress
 	var existingVersion uint16
 	var currentVersion uint16