+++ x/evm/state/accesslist.go
@@ -1,7 +1,6 @@
 package state
 
 import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/params"
@@ -99,8 +98,7 @@ func (s *DBImpl) Prepare(_ params.Rules, sender, coinbase common.Address, dest *
 			if !s.ctx.IsTracing() {
 				continue
 			}
-			upgradeHeight := s.k.UpgradeKeeper().GetDoneHeight(s.ctx.WithGasMeter(sdk.NewInfiniteGasMeterWithMultiplier(s.ctx)), "v5.7.5")
-			if upgradeHeight == 0 || s.ctx.BlockHeight() >= upgradeHeight {
+			if s.ctx.ChainID() != "pacific-1" || s.ctx.BlockHeight() >= 94496767 {
 				continue
 			}
 		}