+++ CHANGELOG.md
@@ -27,6 +27,32 @@ Ref: https://keepachangelog.com/en/1.0.0/
 -->
 
 # Changelog
+## v6.1.0
+sei-chain
+* [#2194](https://github.com/sei-protocol/sei-chain/pull/2194) Fix access list height check
+* [#2187](https://github.com/sei-protocol/sei-chain/pull/2187) Add command to take state sync snapshot at specific height
+* [#2186](https://github.com/sei-protocol/sei-chain/pull/2186) Disable CW -> ERC Register Pointer
+* [#2183](https://github.com/sei-protocol/sei-chain/pull/2183) Add missing methods to distribution precompile
+* [#2180](https://github.com/sei-protocol/sei-chain/pull/2180) Add missing methods to staking precompile
+* [#2179](https://github.com/sei-protocol/sei-chain/pull/2179) Use H+1 oracle for state during tracing
+* [#2176](https://github.com/sei-protocol/sei-chain/pull/2176) Use pointer addr for to address in synthetic tx
+* [#2175](https://github.com/sei-protocol/sei-chain/pull/2175) Update docker with wasm v1.5.5
+* [#2173](https://github.com/sei-protocol/sei-chain/pull/2173) Add missing methods to gov precompile
+* [#2171](https://github.com/sei-protocol/sei-chain/pull/2171) debug_trace Add Timeout + Rate Limit + Lookback + Concurrent calls max
+* [#2166](https://github.com/sei-protocol/sei-chain/pull/2166) Recover panics from unmanaged goroutines
+* [#2163](https://github.com/sei-protocol/sei-chain/pull/2163) Fix gas consumption for historical block tracing
+* [#2158](https://github.com/sei-protocol/sei-chain/pull/2158) Fix oracle extremely slow query
+* [#2156](https://github.com/sei-protocol/sei-chain/pull/2156) Deprecate MinTxsPerBlock
+
+sei-cosmos
+* [#584](https://github.com/sei-protocol/sei-cosmos/pull/584) Add new config OnlyAllowExportOnSnapshotVersion for sc
+* [#580](https://github.com/sei-protocol/sei-cosmos/pull/580) Add nextMs to context
+* [#579](https://github.com/sei-protocol/sei-cosmos/pull/579) Add store tracer
+
+
+sei-tendermint
+* [#284](https://github.com/sei-protocol/sei-tendermint/pull/284) Add godeltapprof to sei-tendermint to serve additional profiling data
+
 ## v6.0.6
 sei-chain
 * [#2161](https://github.com/sei-protocol/sei-chain/pull/2161) Filter EVM Rpc default case